+++ src/Symfony/Bundle/FrameworkBundle/Tests/Controller/ControllerNameParserTest.php
@@ -14,7 +14,6 @@ namespace Symfony\Bundle\FrameworkBundle\Tests\Controller;
 use Composer\Autoload\ClassLoader;
 use Symfony\Bundle\FrameworkBundle\Controller\ControllerNameParser;
 use Symfony\Bundle\FrameworkBundle\Tests\TestCase;
-use Symfony\Component\HttpKernel\Kernel;
 
 /**
  * @group legacy
@@ -106,11 +105,6 @@ class ControllerNameParserTest extends TestCase
             ['FooBundle:Fake:index'],
         ];
 
-        // a bundle with children
-        if (Kernel::VERSION_ID < 40000) {
-            $bundles[] = ['SensioFooBundle:Fake:index'];
-        }
-
         return $bundles;
     }
 
@@ -155,13 +149,13 @@ class ControllerNameParserTest extends TestCase
         $kernel
             ->expects($this->any())
             ->method('getBundle')
-            ->will($this->returnCallback(function ($bundle) use ($bundles) {
+            ->willReturnCallback(function ($bundle) use ($bundles) {
                 if (!isset($bundles[$bundle])) {
                     throw new \InvalidArgumentException(sprintf('Invalid bundle name "%s"', $bundle));
                 }
 
                 return $bundles[$bundle];
-            }))
+            })
         ;
 
         $bundles = [
@@ -172,7 +166,7 @@ class ControllerNameParserTest extends TestCase
         $kernel
             ->expects($this->any())
             ->method('getBundles')
-            ->will($this->returnValue($bundles))
+            ->willReturn($bundles)
         ;
 
         return new ControllerNameParser($kernel);
@@ -181,8 +175,8 @@ class ControllerNameParserTest extends TestCase
     private function getBundle($namespace, $name)
     {
         $bundle = $this->getMockBuilder('Symfony\Component\HttpKernel\Bundle\BundleInterface')->getMock();
-        $bundle->expects($this->any())->method('getName')->will($this->returnValue($name));
-        $bundle->expects($this->any())->method('getNamespace')->will($this->returnValue($namespace));
+        $bundle->expects($this->any())->method('getName')->willReturn($name);
+        $bundle->expects($this->any())->method('getNamespace')->willReturn($namespace);
 
         return $bundle;
     }