+++ src/Symfony/Component/Form/Tests/Extension/Core/Type/TimezoneTypeTest.php
@@ -73,29 +73,6 @@ class TimezoneTypeTest extends BaseTypeTest
     }
 
     /**
-     * @group legacy
-     * @expectedDeprecation The option "regions" is deprecated since Symfony 4.2.
-     */
-    public function testFilterByRegions()
-    {
-        $choices = $this->factory->create(static::TESTED_TYPE, null, ['regions' => \DateTimeZone::EUROPE])
-            ->createView()->vars['choices'];
-
-        $this->assertContains(new ChoiceView('Europe/Amsterdam', 'Europe/Amsterdam', 'Europe / Amsterdam'), $choices, '', false, false);
-    }
-
-    /**
-     * @group legacy
-     * @expectedDeprecation The option "regions" is deprecated since Symfony 4.2.
-     * @expectedException \Symfony\Component\Form\Exception\LogicException
-     * @expectedExceptionMessage The "regions" option can only be used if the "intl" option is set to false.
-     */
-    public function testFilterByRegionsWithIntl()
-    {
-        $this->factory->create(static::TESTED_TYPE, null, ['regions' => \DateTimeZone::EUROPE, 'intl' => true]);
-    }
-
-    /**
      * @requires extension intl
      */
     public function testIntlTimeZoneInput()
@@ -122,6 +99,17 @@ class TimezoneTypeTest extends BaseTypeTest
      */
     public function testIntlTimeZoneInputWithBc()
     {
+        $reflector = new \ReflectionExtension('intl');
+        ob_start();
+        $reflector->info();
+        $output = strip_tags(ob_get_clean());
+        preg_match('/^ICU TZData version (?:=>)?(.*)$/m', $output, $matches);
+        $tzDbVersion = isset($matches[1]) ? (int) trim($matches[1]) : 0;
+
+        if (!$tzDbVersion || 2017 <= $tzDbVersion) {
+            $this->markTestSkipped('"Europe/Saratov" is expired until 2017, current version is '.$tzDbVersion);
+        }
+
         $form = $this->factory->create(static::TESTED_TYPE, null, ['input' => 'intltimezone']);
         $form->submit('Europe/Saratov');
 
@@ -134,6 +122,17 @@ class TimezoneTypeTest extends BaseTypeTest
      */
     public function testIntlTimeZoneInputWithBcAndIntl()
     {
+        $reflector = new \ReflectionExtension('intl');
+        ob_start();
+        $reflector->info();
+        $output = strip_tags(ob_get_clean());
+        preg_match('/^ICU TZData version (?:=>)?(.*)$/m', $output, $matches);
+        $tzDbVersion = isset($matches[1]) ? (int) trim($matches[1]) : 0;
+
+        if (!$tzDbVersion || 2017 <= $tzDbVersion) {
+            $this->markTestSkipped('"Europe/Saratov" is expired until 2017, current version is '.$tzDbVersion);
+        }
+
         $form = $this->factory->create(static::TESTED_TYPE, null, ['input' => 'intltimezone', 'intl' => true]);
         $form->submit('Europe/Saratov');