+++ src/Symfony/Bundle/TwigBundle/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+5.0.0
+-----
+
+ * updated default value for the `strict_variables` option to `%kernel.debug%` parameter
+ * removed support to load templates from the legacy directories `src/Resources/views/` and `src/Resources/<BundleName>/views/`
+
 4.2.0
 -----