+++ src/Symfony/Component/DependencyInjection/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+5.0.0
+-----
+
+* removed support for auto-discovered extension configuration class which does not implement `ConfigurationInterface`
+* removed support for non-string default env() parameters
+
 4.4.0
 -----